+++ besch/bruecke_besch.h
@@ -31,7 +31,6 @@ class bruecke_besch_t : public obj_besch_transport_infrastructure_t {
     friend class bridge_reader_t;
 
 private:
-
 	uint8 pillars_every;	// =0 off
 	bool pillars_asymmetric;	// =0 off else leave one off for north/west slopes
 	uint offset;	// flag, because old bridges had their name/copyright at the wrong position
@@ -110,7 +109,7 @@ public:
 	 * Distance of pillars (=0 for no pillars)
 	 * @author prissi
 	 */
-	int  get_pillar() const { return pillars_every; }
+	uint8  get_pillar() const { return pillars_every; }
 
 	/**
 	 * skips lowest pillar on south/west slopes?
@@ -122,13 +121,13 @@ public:
 	 * maximum bridge span (=0 for infinite)
 	 * @author prissi
 	 */
-	int  get_max_length() const { return max_length; }
+	uint8  get_max_length() const { return max_length; }
 
 	/**
 	 * maximum bridge height (=0 for infinite)
 	 * @author prissi
 	 */
-	int  get_max_height() const { return max_height; }
+	uint8  get_max_height() const { return max_height; }
 
 	void calc_checksum(checksum_t *chk) const;
 };
