+++ besch/skin_besch.h
@@ -26,11 +26,11 @@
  */
 class skin_besch_t : public obj_besch_std_name_t {
 public:
-	bild_besch_t const* get_bild(int i) const { return get_child<bildliste_besch_t>(2)->get_bild(i); }
+	bild_besch_t const* get_bild(uint16 i) const { return get_child<bildliste_besch_t>(2)->get_bild(i); }
 
-	int get_bild_anzahl() const { return get_child<bildliste_besch_t>(2)->get_anzahl(); }
+	uint16 get_bild_anzahl() const { return get_child<bildliste_besch_t>(2)->get_anzahl(); }
 
-	image_id get_bild_nr(int i) const
+	image_id get_bild_nr(uint16 i) const
 	{
 		const bild_besch_t *bild = get_bild(i);
 		return bild != NULL ? bild->get_nummer() : IMG_LEER;
